refactor: Add client-side rendering support to store-backoffice
This commit adds client-side rendering support to the store-backoffice package. The changes include importing the 'use client' statement in the store provider and fetcher hook files. This enables the use of client-side rendering in the application.

The motivation behind this change is to improve the performance and user experience of the backoffice application by offloading some of the rendering tasks to the client side.

Refactor tableTest component to add pagination and sorting functionality
Add tableId prop to TableWarpProps interface
Refactor tableBuilder component to use tableId for sorting and pagination state management
Refactor table component styles, remove unused code, and add state management